Master JWT complete

+++ src/main/java/org/example/azoi/service/impl/TagServiceImpl.java
@@ -4,18 +4,24 @@ import org.example.azoi.dto.Result;
 import org.example.azoi.dto.problemtransmit.othertransmit.TagDTO;
 import org.example.azoi.dto.problemtransmit.othertransmit.TagVO;
 import org.example.azoi.model.problem_model.Tag;
+import org.example.azoi.model.team_model.Role;
+import org.example.azoi.model.user_model.UserRole;
 import org.example.azoi.service.TagService;
 import org.example.azoi.utils.repository.TagRepository;
+import org.example.azoi.utils.repository.UserRoleRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TagServiceImpl implements TagService {
     private final TagRepository tagRepository;
+    private final UserRoleRepository userRoleRepository;
 
-    public TagServiceImpl(TagRepository tagRepository) {
+    public TagServiceImpl(TagRepository tagRepository, UserRoleRepository userRoleRepository) {
         this.tagRepository = tagRepository;
+        this.userRoleRepository = userRoleRepository;
     }
 
     @Override
@@ -36,7 +42,11 @@ public class TagServiceImpl implements TagService {
     }
 
     @Override
-    public Result<Void> addTags(List<TagDTO> tagDTOs) {
+    public Result<Void> addTags(List<TagDTO> tagDTOs, Long requesterId) {
+        Result<Void> result = checkerAdmin(requesterId);
+        if(result.getCode() == Result.FAIL)
+            return new Result<>(null, Result.SUCCESS, result.getMsg());
+
         for(TagDTO tagDTO : tagDTOs)
             addTag(tagDTO);
 
@@ -44,18 +54,42 @@ public class TagServiceImpl implements TagService {
     }
 
     @Override
-    public Result<Void> deleteTag(Long tagId) {
+    public Result<Void> deleteTag(Long tagId, Long requesterId) {
+        Result<Void> result = checkerAdmin(requesterId);
+        if(result.getCode() == Result.FAIL)
+            return new Result<>(null, Result.SUCCESS, result.getMsg());
+
         tagRepository.deleteById(tagId);
         return new Result<>(null, Result.SUCCESS, "ok");
     }
 
     @Override
-    public Result<Void> modifyTag(Tag tag) {
+    public Result<Void> modifyTag(Tag tag, Long requesterId) {
+        Result<Void> result = checkerAdmin(requesterId);
+        if(result.getCode() == Result.FAIL)
+            return new Result<>(null, Result.SUCCESS, result.getMsg());
+
         tagRepository.findById(tag.getId())
                 .ifPresent(tg -> {
                     tg.setName(tag.getName());
                     tg.setColor(tag.getColor());
                 });
+        return new Result<>(null, Result.SUCCESS, "ok");
+    }
+
+    /**
+     * 检查一个角色是否是admin
+     * @param requesterId 请求者Id
+     * @return 实际上Result没有内容，直接查看code
+     */
+    private Result<Void> checkerAdmin(Long requesterId) {
+        //只有管理员才能能添加新的角色
+        Optional<UserRole> requester = userRoleRepository.findById_UserId((requesterId));
+        if (requester.isEmpty())
+            return new Result<>(null, Result.FAIL, "未找到您的信息: requester is empty");
+        if (!requester.get().getRole().getName().equals(Role.ROLE_ADMIN)) {
+            return new Result<>(null, Result.FAIL, "您不是管理员");
+        }
         return new Result<>(null, Result.SUCCESS, "ok");
     }
 }
